Fix numerical problems in sub_pixel_peak()
Sometimes, finding the sub-pixel maximum fails (i.e. the solution is
very far from the maximal value pixel) and the tracked object is lost.

This happens when the quadratic function has no maximum. Therefore, we
use the solution only when we know that the function has a maximum and
also when the maximum lies within one pixel neighbourhood of the
maximal value pixel.