Fix out of bound load and stores
FHoG calculation uses SSE instructions that operate on multiple values
at once. Therefore, depending on the size of the patch, it can happen
that the SSE instructions access (both for reading and writing) the
data outside of the image boundaries. This causes memory corruptions
and crashes.

We fix this problem by allocating more memory so that out-of-image
accesses always touch valid memory.